correct errors from ISOGG database
errors found in the ISOGG 2020 database were found and manually corrected. This affected multiple entries of: FGC65335 (duplicate entry, one of them was the locus for FGC65334). L417 (remove L417 marker for position 8362471 (GRCh37), as this was assigned to another marker. L417 corresponds to another locus. Validated with YBrowse). Z33114 (has 2 different entries; replaced with Z33114.2 for HgM). Y25230 (backmutation, second mutation was labelled Y25230.2). BY115210 (used for 2 separate loci. Replaced with BY115210.2 for hgS. Validated with Yfull and YBrowse). Z27625 (used for 2 separate loci. Replaced with Z27625.2 for hgG. Validated with YFull). PH3539 (used for 2 separate loci. Replaced with PH3539.2 for hgC. Validated with YFull). Z6494 (recurrent mutation for HgG and R. Replaced with Z6494.2 for HgG). PH3797 (used for 2 separate loci. Replaced with PH3797.2 for hgG). Z33633 (used for 2 separate loci. Replaced with Z33633.2 for second entry in hgS). Z26780 (used for 2 separate loci. Replaced with Z26780.2 for entry in hgA).